refactor: 클래스명 및 패키지명 변경
